Cloud-Engineering/Assignments/hw2/423-2024-hw2-scg1143-master/requirements.txt
PyYAML==6.0.1
requests==2.31.0
pandas==2.0.2
matplotlib==3.7.4
boto3==1.34.98
scikit-learn==1.3.2
moto==5.0.6
requests-mock==1.12.1
pylint==3.1.0
pytest==8.2.0
